pyproject.toml
[tool.poetry]
name = "dtf_bot"
version = "0.1.0"
description = ""
authors = ["Evgeny Kemerov <eskemerov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.9"
python-dotenv = "^0.10.1"
psycopg2-binary = "^2.8"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
flake8 = "^3.7"
pylint-django = "^2.0"
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

